Fix SCD4x Input

+++ mycodo/inputs/scd4x_circuitpython.py
@@ -87,7 +87,7 @@ INPUT_INFORMATION = {
         },
         {
             'id': 'altitude',
-            'type': 'float',
+            'type': 'integer',
             'default_value': 0,
             'required': True,
             'name': 'Altitude (m)',
